Created TranslateTTS activity and added a simple layout to work with
