+++ app/src/main/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
-    <uses-permission android:name="android.permission.INTERNET" />
 
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
@@ -13,6 +13,15 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Dialogues"
         tools:targetApi="31">
+        <activity
+            android:name=".TranslationTTS"
+            android:exported="false" />
+        <activity
+            android:name=".Translator"
+            android:exported="false" />
+        <activity
+            android:name=".TTSActivity"
+            android:exported="false" />
         <activity
             android:name=".MainActivity"
             android:exported="true">
